+++ lib/vagrant-templated/command/init.rb
@@ -6,72 +6,65 @@ module Vagrant
         def execute
           template, version, options = setup
           return if template.nil?
-          @env.ui.info("Template detected: #{template}", color: :yellow)
-
-          version = patch template, version
-
-          template_attributes = Catalog.attributes_for template, version
-
-          vagrantfile_save_path = nil
-          berksfile_save_path = nil
-          if options[:suffix] != "-"
-            vagrantfile_save_path = Pathname.new(['Vagrantfile', options[:suffix]].compact.join('.')).expand_path(@env.cwd)
-            vagrantfile_save_path.delete if vagrantfile_save_path.exist? && options[:force]
-            raise Vagrant::Templated::Errors::VagrantfileExistsError if vagrantfile_save_path.exist?
-            berksfile_save_path = Pathname.new(['Berksfile', options[:suffix]].compact.join('.')).expand_path(@env.cwd)
-            berksfile_save_path.delete if berksfile_save_path.exist? && options[:force]
-            raise Vagrant::Templated::Errors::BerksfileExistsError if berksfile_save_path.exist?
+          @env.ui.info "vagrant-templated detected #{template} template", new_line: true
+          if version.nil? || version.empty?
+            version = Catalog.patch template, version
+            @env.ui.info "vagrant-templated didn't received a version, using largest available #{version}"
+          else
+            @env.ui.info "vagrant-templated detected #{version} version"
           end
 
-          vagrantfile = ERB.new File.read(Catalog.root.join 'config/templates/Vagrantfile.erb'), nil, '-'
-          berksfile = ERB.new File.read(Catalog.root.join 'config/templates/Berksfile.erb'), nil, '-'
+          vagrantfile = Catalog.vagrantfile_for template, version
+          berksfile = Catalog.berksfile_for template, version
 
-          contents = vagrantfile.result binding
-          if vagrantfile_save_path
+          if options[:cat]
+            @env.ui.output vagrantfile
+            @env.ui.output berksfile
+          else
+            files_root = @env.cwd
+            if options[:output]
+              @env.ui.info "vagrant-templated received output path #{options[:output]}"
+              output_path = Pathname.new options[:output]
+              files_root =  if output_path.relative?
+                output_path.expand_path files_root
+              else
+                output_path
+              end
+              @env.ui.info "vagrant-templated will write the files in: #{files_root}"
+            end
+
+            vagrantfile_save_path = Pathname.new('Vagrantfile').expand_path files_root
+            vagrantfile_save_path.delete if vagrantfile_save_path.exist? && options[:force]
+            raise Vagrant::Templated::Errors::VagrantfileExistsError if vagrantfile_save_path.exist?
+
+            berksfile_save_path = Pathname.new('Berksfile').expand_path files_root
+            berksfile_save_path.delete if berksfile_save_path.exist? && options[:force]
+            raise Vagrant::Templated::Errors::BerksfileExistsError if berksfile_save_path.exist?
+
             begin
               vagrantfile_save_path.open('w+') do |f|
-                f.write(contents)
+                f.write vagrantfile
               end
+              @env.ui.info "vagrant-templated created #{vagrantfile_save_path}"
             rescue Errno::EACCES
               raise Vagrant::Templated::Errors::VagrantfileWriteError
             end
 
-            @env.ui.info('Templated Vagranfile created successfully', prefix: false)
-          else
-            @env.ui.info('VAGRANTFILE', prefix: false, color: :green)
-            @env.ui.info(contents, prefix: false)
-          end
-
-          contents = berksfile.result binding
-          if berksfile_save_path
             begin
               berksfile_save_path.open('w+') do |f|
-                f.write(contents)
+                f.write berksfile
               end
+              @env.ui.info "vagrant-templated created #{berksfile_save_path}"
             rescue Errno::EACCES
               raise Vagrant::Templated::Errors::BerksfileWriteError
             end
-
-            @env.ui.info('Templated Berksfile created successfully', prefix: false)
-          else
-            @env.ui.info('BERKSFILE', prefix: false, color: :green)
-            @env.ui.info(contents, prefix: false)
           end
-            @env.ui.info('Vagrant Templated finished successfully.', prefix: false, color: :green)
+
+          @env.ui.success 'vagrant-templated finished successfully'
           0
         end
 
         private
-
-          def patch(template, version)
-            if version.nil? || version.empty?
-              version = Catalog.max_version_for template
-              @env.ui.info("No version detected, using: #{version}", color: :yellow)
-            else
-              @env.ui.info("Version detected: #{version}", color: :yellow)
-            end
-            version
-          end
 
           def setup
             options = {}
@@ -91,12 +84,15 @@ module Vagrant
               o.separator 'Options:'
               o.separator ''
 
-              o.on('-s', '--suffix SUFFIX', String,
-                   "Output suffix for Vagrantfile and Berksfile. '-' for stdout") do |suffix|
-                options[:suffix] = suffix
+              o.on('-o', '--output OUTPUT', String, 'Output path for files. Default to `.`') do |o|
+                options[:output] = o
               end
 
-              o.on('-f', '--force', 'Overwrite an existing box if it exists') do |f|
+              o.on('-c', '--cat', 'Output Vagrantfile and Berksfile to stdout') do |c|
+                options[:cat] = c
+              end
+
+              o.on('-f', '--force', 'Overwrite existing Vagrantfile and Berksfile if they exists') do |f|
                 options[:force] = f
               end
             end