+++ CHANGELOG.md
@@ -15,8 +15,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Refactor loading and accessing template attributes
 - Creating all error in vagrant templated module
 - Update vagrantfile and berksfile
+- Replace suffix with output option
 ### Added
 - Node.js 6 template
+- Cat option
 
 
 ## [0.1.5] - 2017-08-07